refactor: 단순 조회 로직과 수정 로직을 분리
