+++ src/main/java/roomescape/business/service/ReservationSlotService.java
@@ -13,22 +13,20 @@ import roomescape.business.model.vo.Id;
 import roomescape.exception.business.NotFoundException;
 
 import java.time.LocalDate;
-import java.util.List;
 
 import static roomescape.exception.ErrorCode.RESERVATION_TIME_NOT_EXIST;
 import static roomescape.exception.ErrorCode.THEME_NOT_EXIST;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 public class ReservationSlotService {
 
     private final ReservationSlots reservationSlots;
     private final Themes themes;
     private final ReservationTimes reservationTimes;
 
-    @Transactional
-    public ReservationSlot saveAndGet(final LocalDate date, final String reservationTimeIdValue, final String themeIdValue) {
+    public ReservationSlot addAndGet(final LocalDate date, final String reservationTimeIdValue, final String themeIdValue) {
         Theme theme = themes.findById(Id.create(themeIdValue))
                 .orElseThrow(() -> new NotFoundException(THEME_NOT_EXIST));
         ReservationTime time = reservationTimes.findById(Id.create(reservationTimeIdValue))
@@ -37,14 +35,5 @@ public class ReservationSlotService {
         ReservationSlot createdSlot = new ReservationSlot(time, date, theme);
         reservationSlots.save(createdSlot);
         return createdSlot;
-    }
-
-    public ReservationSlot getByDateAndTimeIdAndThemeIdOrElseCreate(final LocalDate date, final String reservationTimeIdValue, final String themeIdValue) {
-        return reservationSlots.findByDateAndTimeIdAndThemeId(date, Id.create(reservationTimeIdValue), Id.create(themeIdValue))
-                .orElseGet(() -> saveAndGet(date, reservationTimeIdValue, themeIdValue));
-    }
-
-    public List<ReservationSlot> getAllSlotsContainsReserverOf(final String userIdValue) {
-        return reservationSlots.findAllSlotsContainsReserverOf(Id.create(userIdValue));
     }
 }
