+++ src/main/java/roomescape/business/service/ReservationTimeService.java
@@ -3,7 +3,6 @@ package roomescape.business.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.business.dto.ReservableReservationTimeDto;
 import roomescape.business.dto.ReservationTimeDto;
 import roomescape.business.model.entity.ReservationTime;
 import roomescape.business.model.repository.ReservationTimes;
@@ -14,9 +13,7 @@ import roomescape.exception.business.InvalidCreateArgumentException;
 import roomescape.exception.business.NotFoundException;
 import roomescape.exception.business.RelatedEntityExistException;
 
-import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.List;
 
 import static roomescape.exception.ErrorCode.RESERVATION_NOT_EXIST;
 import static roomescape.exception.ErrorCode.RESERVATION_TIME_ALREADY_EXIST;
@@ -25,13 +22,12 @@ import static roomescape.exception.ErrorCode.RESERVED_RESERVATION_TIME;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 public class ReservationTimeService {
 
     private final ReservationTimes reservationTimes;
     private final Reservations reservations;
 
-    @Transactional
     public ReservationTimeDto addAndGet(final LocalTime time) {
         ReservationTime reservationTime = new ReservationTime(time);
         validateNoDuplication(reservationTime);
@@ -55,20 +51,6 @@ public class ReservationTimeService {
         }
     }
 
-    public List<ReservationTimeDto> getAll() {
-        List<ReservationTime> reservationTimes = this.reservationTimes.findAll();
-        return ReservationTimeDto.fromEntities(reservationTimes);
-    }
-
-    public List<ReservableReservationTimeDto> getAllByDateAndThemeId(final LocalDate date, final String themeIdValue) {
-        Id themeId = Id.create(themeIdValue);
-        List<ReservationTime> available = reservationTimes.findAvailableByDateAndThemeId(date, themeId);
-        List<ReservationTime> notAvailable = reservationTimes.findNotAvailableByDateAndThemeId(date, themeId);
-
-        return ReservableReservationTimeDto.fromEntities(available, notAvailable);
-    }
-
-    @Transactional
     public void delete(final String timeIdValue) {
         Id timeId = Id.create(timeIdValue);
         if (reservations.existByTimeId(timeId)) {
