+++ src/main/java/roomescape/presentation/api/UserApiController.java
@@ -10,6 +10,7 @@ import roomescape.auth.AuthRequired;
 import roomescape.auth.Role;
 import roomescape.business.dto.UserDto;
 import roomescape.business.model.vo.UserRole;
+import roomescape.business.reader.UserReader;
 import roomescape.business.service.UserService;
 import roomescape.presentation.dto.request.RegisterRequest;
 import roomescape.presentation.dto.response.UserResponse;
@@ -22,12 +23,13 @@ import java.util.List;
 public class UserApiController {
 
     private final UserService userService;
+    private final UserReader userReader;
 
     @GetMapping("/members")
     @AuthRequired
     @Role(UserRole.ADMIN)
     public ResponseEntity<List<UserResponse>> getUsers() {
-        List<UserDto> users = userService.getAll();
+        List<UserDto> users = userReader.getAll();
         List<UserResponse> responses = UserResponse.from(users);
         return ResponseEntity.ok(responses);
     }
